3 Modes v1
Added weapon resistance mode.
Removed max pixel display.
Should work properly, but may have issues with Serial output.

+++ FencingTestBox_v0.3/eeprom.ino
@@ -14,7 +14,7 @@ void setEepromLocation() {
 }
 
 void writeCalibrationData() {
-  int loc;
+  uint16_t loc;
 
   if ((EEPROM.read(eepromAddr)==calibrationValid)) {
     EEPROM.update(eepromAddr,calibrationInvalid); //Mark old calibration as invalid
